requirements.txt
psycopg2-binary>=2.9
requests>=2.32

